requirements.txt
Flask
python-dotenv
PyPDF2
python-pptx
python-docx
pandas
sentence-transformers
faiss-cpu
transformers
torch